# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApiResponse": {
        "properties": {
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "description": "业务数据",
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ],
            "description": "错误信息"
          },
          "success": {
            "default": true,
            "description": "请求是否成功",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "ApiResponse",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "错误代码",
            "title": "Code",
            "type": "string"
          },
          "message": {
            "description": "错误说明",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "GatewayHeartbeatRequest": {
        "properties": {
          "current_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Run Id"
          },
          "metrics": {
            "additionalProperties": true,
            "title": "Metrics",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "GatewayHeartbeatRequest",
        "type": "object"
      },
      "GatewayRegisterRequest": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "agent_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Version"
          },
          "capabilities": {
            "additionalProperties": true,
            "title": "Capabilities",
            "type": "object"
          },
          "gateway_id": {
            "title": "Gateway Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "gateway_id",
          "name"
        ],
        "title": "GatewayRegisterRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "SimChip Nexus 芯境智测平台提供 CARLA 场景仿真、硬件在环测试、设备接入和评测报告服务。\n\n- `/docs`：接口文档\n- `/`：平台展示页\n- `/ui`：Web 测评控制台",
    "title": "SimChip Nexus API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/gateways/register": {
      "post": {
        "operationId": "register_gateway_gateways_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewayRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "注册网关",
        "tags": [
          "网关管理"
        ]
      }
    },
    "/gateways/{gateway_id}": {
      "get": {
        "operationId": "get_gateway_gateways__gateway_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "gateway_id",
            "required": true,
            "schema": {
              "title": "Gateway Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询单个网关",
        "tags": [
          "网关管理"
        ]
      }
    },
    "/gateways/{gateway_id}/heartbeat": {
      "post": {
        "operationId": "heartbeat_gateway_gateways__gateway_id__heartbeat_post",
        "parameters": [
          {
            "in": "path",
            "name": "gateway_id",
            "required": true,
            "schema": {
              "title": "Gateway Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewayHeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "网关心跳",
        "tags": [
          "网关管理"
        ]
      }
    },
...
    "/projects/{project_id}": {
      "get": {
        "operationId": "get_project_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询单个项目",
        "tags": [
          "项目管理"
        ]
      }
    },
    "/projects/{project_id}/workspace": {
      "get": {
        "operationId": "get_project_workspace_projects__project_id__workspace_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询项目工作台数据",
        "tags": [
          "项目管理"
        ]
      }
    },
...
  }
}